Create Karaoke Song Selector
Ever sung karaoke? It's a place where you sing songs with your friends, and it's a lot of fun. I've created a table with songs, and in this challenge, I used queries to decide what songs to sing. For the first step, I selected all the song titles. However, maybe my friends only like singing either recent songs or truly epic songs. I added another SELECT that uses OR to show the titles of the songs that have an 'epic' mood or a release date after 1990. Now, people get picky at the end of the night. I also added another SELECT that uses AND to show the titles of songs that are 'epic', and released after 1990, and less than 4 minutes long.