Create Playlist Maker
I've created a database of songs and artists, and I've made playlists from them in this challenge. 

In this first step, I've selected the title of all the songs by the artist named 'Queen'.

Now I've made a 'Pop' playlist.

To finish creating the 'Pop' playlist, I've added another query that will select the title of all the songs from the 'Pop' artists.